Various
Fixed a bug that meant the hot water request did not turn off unless setpoint was met while the heating programmer hot water enable was on. It now turnes the hot water request off as soon as the programmer enable is false.

Tidied the code a little bit to remove some of the repetition, while still a rapid prototype it's now a little bit more respectable. Ideally this needs a rewrite with unit tests.

Added the ability to set the hot water set point via a new CAN packet. This will allow control via a new programmer node when I've got time to build one.